Various fixes to examples and update tenant (#251)
* Fix search example event formatting and use main index.

* Change examples host and source to epoch seconds to avoid collisions.

* Fix metrics example SPL and run all examples even if they fail.

* Don't delete any datasets that are createdby Splunk.

* Update tenant to one with passthru pipeline.

* Fix issue with unhandled search promise.

* Revise metric SPL.

* Add more logging to examples.

+++ test/integration/catalog_proxy.js
@@ -211,8 +211,8 @@ function deleteAllDatasets() {
             .then(datasets => {
                 return Promise.all(
                     datasets.map(dataset => {
-                        // Delete the dataset unless it's a main/metrics
-                        if (dataset.name !== "main" && dataset.name !== "metrics") {
+                        // Delete the dataset unless it's a main/metrics or createdby Splunk
+                        if (dataset.name !== "main" && dataset.name !== "metrics" && dataset.createdby !== "Splunk") {
                             return splunkCloud.catalog.deleteDataset(dataset.id);
                         }
                     })
